refactor direct select display helpers to separate file
