+++ src/DashboardUI/src/components/map/CustomDirectSelectMode/CustomDirectSelectMode.tsx
@@ -1,8 +1,9 @@
 import MapboxDraw, { DrawCustomMode, DrawCustomModeThis } from '@mapbox/mapbox-gl-draw';
 import center from '@turf/center';
 import { LngLat } from 'mapbox-gl';
-import { Feature, GeoJSON, GeoJsonProperties, Geometry, Position } from 'geojson';
+import { Feature, GeoJSON, GeoJsonProperties, Geometry } from 'geojson';
 import { dragFeature, dragVertex } from './CustomDirectSelectModeDragHelpers';
+import { handleDisplayCircle, handleDisplayRectangle } from './CustomDirectSelectModeDisplayHelpers';
 
 // base code used for reference: https://github.com/mapbox/mapbox-gl-draw/blob/main/src/modes/direct_select.js
 
@@ -77,60 +78,11 @@ export const CustomDirectSelectMode: DrawCustomMode = {
     geojson: Feature<Geometry, GeoJsonProperties>,
     display: (geojson: GeoJSON) => void,
   ) {
-    const getCardinalDirectionCoordinatesOnFeature = (): Position[] => {
-      // technically this gets the coordinates that align *most closely* with the cardinal directions
-      // but the distinction doesn't really matter for a circle containing a sufficient number of points
-      const allCoordinateLngLatValues = state.feature!.getCoordinates()[0];
-      const north = allCoordinateLngLatValues.reduce((prevCoord, currentCoord) =>
-        prevCoord[1] > currentCoord[1] ? prevCoord : currentCoord,
-      );
-      const east = allCoordinateLngLatValues.reduce((prevCoord, currentCoord) =>
-        prevCoord[0] > currentCoord[0] ? prevCoord : currentCoord,
-      );
-      const south = allCoordinateLngLatValues.reduce((prevCoord, currentCoord) =>
-        prevCoord[1] < currentCoord[1] ? prevCoord : currentCoord,
-      );
-      const west = allCoordinateLngLatValues.reduce((prevCoord, currentCoord) =>
-        prevCoord[0] < currentCoord[0] ? prevCoord : currentCoord,
-      );
-      return [north, east, south, west];
-    };
-
-    const buildVertex = (parentId: string, coordinates: Position): Feature<Geometry, GeoJsonProperties> => ({
-      type: 'Feature',
-      properties: {
-        meta: 'vertex',
-        parent: parentId,
-        coord_path: '0.0',
-        active: 'true',
-      },
-      geometry: {
-        type: 'Point',
-        coordinates: [coordinates[0], coordinates[1]],
-      },
-    });
-
     // determine whether to override the base implementation
     if (state.feature?.properties?.isCircle && state.featureId === geojson.properties?.id) {
-      // display the circle feature
-      geojson.properties.active = 'true';
-      display(geojson);
-
-      // render marker points manually rather than add them to the map's feature collection
-      const draggableMarkerPositions = getCardinalDirectionCoordinatesOnFeature();
-      draggableMarkerPositions
-        .map((position) => buildVertex(geojson.properties!.id, position))
-        .forEach((vertex) => display(vertex));
+      handleDisplayCircle(state, geojson, display);
     } else if (state.feature?.properties?.isRectangle && state.featureId === geojson.properties?.id) {
-      // override rendering - default polygon tool displays midpoint markers which we don't want
-      // display the rectangle feature
-      geojson.properties.active = 'true';
-      display(geojson);
-
-      // render marker points manually rather than add them to the map's feature collection
-      const rectangle = state.feature;
-      const corners = rectangle.getCoordinates()[0];
-      corners.map((position) => buildVertex(geojson.properties!.id, position)).forEach((vertex) => display(vertex));
+      handleDisplayRectangle(state, geojson, display);
     } else {
       // call base implementation
       directSelectBaseMode.toDisplayFeatures?.call(this, state, geojson, display);
